requirements.txt
Flask==2.2.3
gunicorn==20.1.0
google-cloud-bigquery==3.5.0
jsonschema==4.17.3
python-dotenv==0.21.1